Improve home screen by removing useless name print
