+++ src/components/Home/Balance.js
@@ -7,31 +7,15 @@
  */
 
 import React from 'react';
-import { Text, View } from 'react-native';
-import Ionicons from '@expo/vector-icons/Ionicons';
+import { Text } from 'react-native';
 import BlockTemplate from '../BlockTemplate';
 import { Home as t } from '../../utils/i18n';
 import colors from '../../styles/colors';
 import { floatToEuro } from '../../utils';
 
-const shortcuts = [
-	{
-		screen: 'Refill',
-		lazyTitle: 'refill',
-		icon: 'ios-add-circle-outline',
-		color: colors.more,
-	},
-	{
-		screen: 'Transfer',
-		lazyTitle: 'transfer',
-		icon: 'ios-share-alt',
-		color: colors.lightBlue,
-	},
-];
-
 export default class Balance extends React.PureComponent {
 	render() {
-		const { amount, navigation } = this.props;
+		const { amount } = this.props;
 
 		return (
 			<BlockTemplate roundedTop roundedBottom shadow>
@@ -47,31 +31,6 @@ export default class Balance extends React.PureComponent {
 						{t('no_balance')}
 					</Text>
 				)}
-				<View style={{ flexDirection: 'row', justifyContent: 'space-around', paddingTop: 5 }}>
-					{shortcuts.map(({ screen, lazyTitle, icon, color }) => (
-						<BlockTemplate
-							roundedTop
-							roundedBottom
-							shadow
-							key={lazyTitle}
-							onPress={() => navigation.navigate(screen, { credit: amount })}
-						>
-							<View style={{ flexDirection: 'row', alignItems: 'center', paddingHorizontal: 10 }}>
-								<Ionicons name={icon} size={17} style={{ color }} />
-								<Text
-									style={{
-										paddingLeft: 5,
-										fontSize: 15,
-										fontWeight: 'bold',
-										color,
-									}}
-								>
-									{t(lazyTitle)}
-								</Text>
-							</View>
-						</BlockTemplate>
-					))}
-				</View>
 			</BlockTemplate>
 		);
 	}
